- Move MapEnumerator to a separate file. - Add String8.ToUpperInvariant *in place*. - Fix String8Block.GetBlockForLength bug (allocated new block AND reused existing one when length too long). - Make String8.Concatenate not horribly wasteful when final value is over 64k.
String8.Concatenate would concatenate in place (good), until requiring a separate buffer. Then, each concat allocated another exact size buffer and kept all previous ones (very bad).

Now, String8.Concatenate will allocate an oversized buffer when it needs to allocate, so future small concats can be in place again. Also, if the concat value is moved off of the end of an existing block, it is "removed" so the block can be reused. This means the "current" block is freed up to be used by the next concat when the first one gets big enough to require separate space.