Implemented ItemMap merging.   - MapEnumerator: Exposing CurrentIndex.
Merging allows data structures with maps to round trip multiple times. I needed this because I needed to search for items to add links to them. I add the items themselves, ConvertToImmutable, find them and add the links, and Convert again to get a final converted map.

This shows an efficient model for merging - copy from the existing set and look for added values in between the existing ones. A less efficient alternative would be to add all immutable items to the mutable set and then convert it. In the ItemMap case, I have maps with millions of links which would be expensive to reconvert this way.