Fix Elfie Unit Tests.
So, upgrading Newtonsoft.Json to 10.0.1 caused unit test failures. Specifically, you get a MissingMethodException for the CSharpCompilationOptions constructor. I *think* this is because somehow Newtonsoft.Json uses a different version of System.Collections.Immutable (there was also a build warning about the System.Collections.Immutable version redirect in app.config being wrong). A System.Collections.Immutable type is used in the constructor, so I guess it won't bind anymore.

I'm changing back to Newtonsoft.Json 9.0.1. It appears that when we update NuGet packages we'll need to update them all together, so the recursive dependencies are all for matching versions.