Remove some additional allocations from the insert path (#133)
On every insert, a new typed worker was being created for every column of every partition. This is made even more expensive because of the reflection work needed to build the generic type at runtime. By caching the worker object as well as the value object within the worker, we can eliminate allocations after the first insert operation.