XForm: Timeout via CancellationToken.   - Make IXTable.Next take a CancellationToken.   - Underlying readers stop reading if cancelled; other verbs ignore.   - Reimplement RunUntilTimeout via CancellationToken.   - Add ValidatingTable check for token being passed up.   - Add ValidatingTable case for every verb.
'RunUntilTimeout' didn't work correctly before because it would only stop when another non-zero batch of rows was returned from the whole query. If a query doesn't match anything, it'll run completely without checking for timeout at all.

CancellationTokens are now passed through and the underlying readers check it before each batch, so we now check for cancellation before reading each batch of 10k rows from the source at the top of the query.